# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC Online Draft System",
    "description": "Realtime multi-team player draft system for Vietnamese FC Online pro tournament.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/player-seasons/{card_id}": {
      "get": {
        "tags": [
          "Player Seasons"
        ],
        "summary": "Get Player Season",
        "description": "Get player season card detail by UUID.",
        "operationId": "get_player_season_api_player_seasons__card_id__get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSeasonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerEmbedded": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "externalPlayerId": {
            "type": "string",
            "title": "Externalplayerid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "externalPlayerId"
        ],
        "title": "PlayerEmbedded"
      },
...
      "PlayerSeasonResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "playerId": {
            "type": "string",
            "format": "uuid",
            "title": "Playerid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid",
            "title": "Seasonid"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "salary": {
            "type": "integer",
            "title": "Salary"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerEmbedded"
          },
          "season": {
            "$ref": "#/components/schemas/SeasonEmbedded"
          }
        },
        "type": "object",
        "required": [
          "id",
          "playerId",
          "seasonId",
          "position",
          "salary",
          "status",
          "player",
          "season"
        ],
        "title": "PlayerSeasonResponse"
      },
...
      "SeasonEmbedded": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "badgeUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Badgeurl"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code"
        ],
        "title": "SeasonEmbedded"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}